Moved frustum culling debug info, and added frame synchronization profiling to gkDebugFps.
Enabled simple static geometry again. 
For static geometry I added a special entity unload method, so physics 
and game logic can still do tests on the object.

Added a test for blender relative file paths, this was causing a lag when 
testing for file existence.

Added the ability to reset logic sensors to the initial creation state 
when the scene is reloaded.

Updated sound manager to use distance model settings from the .blend

Updated the main test .blend to use a radar sensor for testing 
if the player is on the ground.


git-svn-id: http://gamekit.googlecode.com/svn/trunk@484 c957766c-2088-11de-b650-e715bd6d7cf1