Finish the simple object animation system started by Charlie. Each action have now a (facultative)  gkGameObjectChannel in additon to the (facultative) gkActionChannel list (one per bone). gkGameObject have a list of actions (containing only object animation) while gkEntity still keeps his lits of actions too (containing only bone animations). There might be more abstraction on the gkAction level as in the end an action will have only an object channel or (several) bone channel but never both.
The gkActionActuator logic brick was updated to play objects action too.
The inconsistenciesin ways of playing skeleton actions and objects action are visible there but still not mess the code too much.

The loader convert Blender IPO actuator to gkActionActuator but blender actuator does not specify whixh action to pay and alwasy play the one active in the UI. I still did not find a way to figure how to retrieve this action so for now all IPO actuatores are set to play the action "action" when converted.
The loader load Blender >2.5 only and works only for quaternion f-curves.

Expect a loader for 2.4 that figure out the best action to use soon. Certainly some cleaning;refactoring once I will have some feedback/better idea on how to integrate it better.



git-svn-id: http://gamekit.googlecode.com/svn/trunk@697 c957766c-2088-11de-b650-e715bd6d7cf1