from now on every scene has its own logic(brick)manager. the singleton maps to the logicmanager that is owned by the activescene.
Since we got dithi's addOverlayScene and such we are able to have
multiple scenes visible at once. in the update-method of gkScene the
logicmanager-singleton's update was called. Which had the consequence
that if you got n scenes visible the logicmanager's update method was
called n-times per tick. 
Now every scene has its own logic(brick)manager which can be accessed
via ->getLogicBrickManager()
The former way to access the logicmanager via singleton should be
avoided. I kept this method due to compatibility reasons. At the
moment the singleton returns the logicmanager of the active scene.

lua-access kept more or less the same. When you create a LogicManager
via OgreKit.LogicManager() you would create a logic manager for the
activescene. New bricks would be execute in activescene's update-method.
If you want to add bricks in other scenes you can query any scene via
OgreKit.getScene("name") and pass this one to the logicmanager: 
e.g. 
scene2=OgreKit.getScene("Scene.002")
logicmanager = OgreKit.LogicManager(scene2)

There are two new regression-files TestLuaLogicBricks.blend (which is
basically the same sample as shown in the lua-documentation) and
TestMultipleSceneLogicbrickManagers (which demonstrates to suspend one
scene and its logicbricks and to let the other scene go on)

git-svn-id: http://gamekit.googlecode.com/svn/trunk@1136 c957766c-2088-11de-b650-e715bd6d7cf1